+++ src/monitors/cycle/registerSuccess.ts
@@ -1,18 +1,20 @@
 import { MonitorType, UniqueMonitorConfig } from '@/config/Config'
 import logger from '@/logger'
-import { registerSuccess } from '../statusRegistry'
+import { registerSuccess, SuccessfulCheckRegistrator } from '../statusRegistry'
 
 /** Handles cycle success */
 export type CycleSuccessRegistrator = <T extends MonitorType>(uniqueConfig: UniqueMonitorConfig<T>) => Promise<void>
 
 const log = (id: string) => logger.debug('Monitor cycle success', { id });
 
-const fn: CycleSuccessRegistrator = (uniqueConfig) => {
-  log(uniqueConfig.id);
+export const base = 
+  (updateStatus: SuccessfulCheckRegistrator): CycleSuccessRegistrator =>
+  (uniqueConfig) => {
+    log(uniqueConfig.id);
 
-  registerSuccess(uniqueConfig.id);
+    updateStatus(uniqueConfig.id);
 
-  return Promise.resolve();
-};
+    return Promise.resolve();
+  }
 
-export default fn
+export default base(registerSuccess)