+++ src/notifications/shouldNotifyError.ts
@@ -5,7 +5,7 @@ export type ErrorNotificationFilter = (monitorId: string, error: Error) => Promi
 export const base = 
   (isRecurringError: RecurringErrorChecker): ErrorNotificationFilter =>
   (monitorId: string, error: Error) => 
-    Promise.resolve(isRecurringError(monitorId, error));
+    Promise.resolve( !isRecurringError(monitorId, error) );
 
 /** Checks if error is repeating in sequence without being resolved */
 export default base(isRecurringError)