+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+## [1.0.1-beta.4](https://github.com/marius321967/web-monitor/compare/v1.0.1-beta.3...v1.0.1-beta.4) (2022-11-27)
+
+
+### Bug Fixes
+
+* auth_header in monitor request is optional ([0c6e3d5](https://github.com/marius321967/web-monitor/commit/0c6e3d51630b6af329ae167144ec0b36cc240d24))
+
+## [1.0.1-beta.3](https://github.com/marius321967/web-monitor/compare/v1.0.1-beta.2...v1.0.1-beta.3) (2022-11-26)
+
+
+### Bug Fixes
+
+* **notifications:** fix repeating notifications ([1b18a87](https://github.com/marius321967/web-monitor/commit/1b18a8727ed892883e7e8cacff0c0b672e585862))
+
+## [1.0.1-beta.2](https://github.com/marius321967/web-monitor/compare/v1.0.1-beta.1...v1.0.1-beta.2) (2022-11-23)
+
+
+### Bug Fixes
+
+* **docker:** point default command to the correct file ([8fae552](https://github.com/marius321967/web-monitor/commit/8fae5528622bfeaaf6eb591e43a045f3a0b46ca0))
+
+## [1.0.1-beta.1](https://github.com/marius321967/web-monitor/compare/v1.0.0...v1.0.1-beta.1) (2022-11-23)
+
+
+### Bug Fixes
+
+* **port:** drop port expose from dockerfile ([efe48df](https://github.com/marius321967/web-monitor/commit/efe48df6913f1f4cfb6f37a209857d048c23055a))
+
 # 1.0.0 (2022-11-20)